Refacttor
